requirements.txt
# requirements.txt (Dependencies)
streamlit>=1.32.0
groq>=0.3.0
bcrypt>=4.0.1
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.1
langchain-chroma>=0.1.0
langchain-huggingface
pdfplumber>=0.10.0
chromadb>=0.4.0
sentence-transformers